refactor to cleanup __init__
