+++ hvplot/tests/testhelp.py
@@ -4,6 +4,8 @@ import pytest
 from holoviews.core import Store
 from holoviews.element import Curve
 
+from hvplot.util import _get_doc_and_signature
+
 
 @pytest.fixture
 def reset_default_backend():
@@ -14,7 +16,7 @@ def reset_default_backend():
 
 def test_help_style_extension_output(reset_default_backend):
     # default, after e.g. import hvplot.pandas
-    docstring, signature = hvplot._get_doc_and_signature(
+    docstring, signature = _get_doc_and_signature(
         cls=hvplot.hvPlot,
         kind='line',
         completions=False,
@@ -29,7 +31,7 @@ def test_help_style_extension_output(reset_default_backend):
 
     # The current backend becomes matplotlib
     hvplot.extension('matplotlib', 'plotly')
-    docstring, signature = hvplot._get_doc_and_signature(
+    docstring, signature = _get_doc_and_signature(
         cls=hvplot.hvPlot,
         kind='line',
         completions=False,
@@ -44,7 +46,7 @@ def test_help_style_extension_output(reset_default_backend):
 
     # The current backend becomes plotly
     hvplot.output(backend='plotly')
-    docstring, signature = hvplot._get_doc_and_signature(
+    docstring, signature = _get_doc_and_signature(
         cls=hvplot.hvPlot,
         kind='line',
         completions=False,
@@ -61,7 +63,7 @@ def test_help_style_extension_output(reset_default_backend):
 def test_help_style_compatibility(reset_default_backend):
     # The current backend is plotly but the style options are those of matplotlib
     hvplot.extension('plotly', 'matplotlib', compatibility='matplotlib')
-    docstring, signature = hvplot._get_doc_and_signature(
+    docstring, signature = _get_doc_and_signature(
         cls=hvplot.hvPlot,
         kind='line',
         completions=False,
